Move default credentials setup to inside FunctionsEmulator
